+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2024-06-20
+
+- Added a plugin counter "badge" to the plugins page button for when new plugins
+  are added as part of a mod installation.
+- Added informational popover to explain how LOOT works.
+- Added LOOT message for plugins which are not present in the masterlist
+
 ## [0.2.9] - 2024-03-14
 
 - Fixed long delays when ascertaining if a plugin is marked light